refactor(amqp_client): move amqp_client to lib
amqp_client.js was missed in the "great cleanup" effort because it
was residing in the top level directory. I've moved it to the lib
folder and pointed to the new location in the package.json. As an
added bonus I also modified all the examples to actually point to
valid locations now, though I have not added them to the jshint/lint
steps as of yet.
