+++ lib/types.js
@@ -476,16 +476,16 @@ Types.prototype._initTypesArray = function() {
           'label': '64-bit unsigned integer in network byte order',
           'builder': function(val, bufb) {
             if (val instanceof Int64) {
-                            bufb.appendBuffer(val.toBuffer(true));
+              bufb.appendBuffer(val.toBuffer(true));
             } else if (typeof val === 'number') {
-                            if (val < 0xFFFFFFFF) {
+              if (val < 0xFFFFFFFF) {
                 bufb.appendUInt32BE(0x0);
                 bufb.appendUInt32BE(val);
-                            } else {
+              } else {
                 throw new exceptions.NotImplementedError('No int64 Number supported by buffer builder');
-                            }
+              }
             } else {
-                            throw new Error('Invalid encoding type for 64-bit value: ' + val);
+              throw new Error('Invalid encoding type for 64-bit value: ' + val);
             }
           },
           'decoder': function(buf) { return new Int64(buf); }
@@ -589,17 +589,17 @@ Types.prototype._initTypesArray = function() {
           'label': "64-bit two's-complement integer in network byte order",
           'builder': function(val, bufb) {
             if (val instanceof Int64) {
-                            bufb.appendBuffer(val.toBuffer(true));
+              bufb.appendBuffer(val.toBuffer(true));
             } else if (typeof val === 'number') {
-                            var absval = Math.abs(val);
-                            if (absval < 0xFFFFFFFF) {
+              var absval = Math.abs(val);
+              if (absval < 0xFFFFFFFF) {
                 bufb.appendUInt32BE(val < 0 ? 0xFFFFFFFF : 0x0);
                 bufb.appendUInt32BE(val < 0 ? (0xFFFFFFFF - absval + 1) : absval);
-                            } else {
+              } else {
                 throw new exceptions.NotImplementedError('buffer-builder does not support 64-bit int appending');
-                            }
+              }
             } else {
-                            throw new Error('Invalid encoding type for 64-bit value: ' + val);
+              throw new Error('Invalid encoding type for 64-bit value: ' + val);
             }
           },
           'decoder': function(buf) { return new Int64(buf); }
@@ -756,17 +756,17 @@ Types.prototype._initTypesArray = function() {
           'label': '64-bit signed integer representing milliseconds since the unix epoch',
           'builder': function(val, bufb) {
             if (val instanceof Int64) {
-                            bufb.appendBuffer(val.toBuffer(true));
+              bufb.appendBuffer(val.toBuffer(true));
             } else if (typeof val === 'number') {
-                            var absval = Math.abs(val);
-                            if (absval < 0xFFFFFFFF) {
+              var absval = Math.abs(val);
+              if (absval < 0xFFFFFFFF) {
                 bufb.appendUInt32BE(val < 0 ? 0xFFFFFFFF : 0x0);
                 bufb.appendUInt32BE(val < 0 ? (0xFFFFFFFF - absval + 1) : absval);
-                            } else {
+              } else {
                 throw new exceptions.NotImplementedError('buffer-builder does not support 64-bit int appending');
-                            }
+              }
             } else {
-                            throw new Error('Invalid encoding type for 64-bit value: ' + val);
+              throw new Error('Invalid encoding type for 64-bit value: ' + val);
             }
           },
           'decoder': function(buf) { return new Int64(buf); }
