+++ lib/types/amqp_symbol.js
@@ -1,5 +1,7 @@
 'use strict';
 
+
+
 /**
  * Encoding for AMQP Symbol type, to differentiate from strings.  More terse than ForcedType.
  *
