+++ lib/types/forced_type.js
@@ -1,5 +1,7 @@
 'use strict';
 
+
+
 /**
  * ForcedType coerces the encoder to encode to the given type, regardless of what it might think.
  *
