+++ package.json
@@ -2,7 +2,7 @@
   "name": "node-amqp-1-0",
   "version": "0.0.15",
   "description": "AMQP-1.0 compatible Node client, not based on native code (e.g. no proton).",
-  "main": "amqp_client.js",
+  "main": "./lib/amqp_client.js",
   "engines": {
     "node": ">= 0.8"
   },
